Deleted GameState class (methods were relocated)
